pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "art-activity-collection"
version = "0.1.0"
description = "Collect and serve free youth art activities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "pymysql>=1.1.1",
  "PyJWT[crypto]>=2.9.0",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
crawler = [
  "playwright>=1.46.0",
  "tenacity>=8.5.0",
]
llm = [
  "openai>=1.40.0",
]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
